Update sketch.js
At first, I wanted to create a game-like narration where the user could click on the raw bacon image to cook it but I couldn't figure out how to use the mouseReleased or mousePressed function in p5 so I decided to switch to a time based narration. Trying to figure out the order of the if statements was difficult. I was originally trying to have the raw bacon move while the pan was moving (its commented out in my code) but after a couple tries, it still wasn't working so I had to keep the bacon still. But, I still got the pan to work and the bacon and egg to move to the correct location!